1/consumer/requirements.txt
nats-py
asyncio